+++ lensless/eval/benchmark.py
@@ -9,10 +9,12 @@
 
 import glob
 import os
-from lensless.io import load_psf, load_image
-from lensless.util import resize
+from lensless.utils.io import load_psf
+from lensless.utils.image import resize
 import numpy as np
 from tqdm import tqdm
+
+from lensless.utils.io import load_image
 
 try:
     import torch
@@ -20,9 +22,10 @@ try:
     from torch.nn import MSELoss, L1Loss
     from torchmetrics import StructuralSimilarityIndexMeasure
     from torchmetrics.image import lpip, psnr
-    from torchvision import transforms
 except ImportError:
-    raise ImportError("Torch and torchmetrics are needed to benchmark reconstruction algorithm")
+    raise ImportError(
+        "Torch, torchvision, and torchmetrics are needed to benchmark reconstruction algorithm."
+    )
 
 
 class ParallelDataset(Dataset):
@@ -206,6 +209,8 @@ class DiffuserCamTestDataset(ParallelDataset):
         )
 
         # transform from BGR to RGB
+        from torchvision import transforms
+
         transform_BRG2RGB = transforms.Lambda(lambda x: x[..., [2, 1, 0]])
 
         self.psf = transform_BRG2RGB(torch.from_numpy(psf))