+++ lensless/utils/image.py
@@ -1,5 +1,5 @@
 # #############################################################################
-# util.py
+# image_utils.py
 # =================
 # Authors :
 # Eric BEZZAM [ebezzam@gmail.com]
@@ -8,12 +8,8 @@
 
 
 import cv2
-import csv
 import numpy as np
-import paramiko
-from paramiko.ssh_exception import BadHostKeyException, AuthenticationException, SSHException
-import socket
-from lensless.constants import RPI_HQ_CAMERA_CCM_MATRIX, RPI_HQ_CAMERA_BLACK_LEVEL
+from lensless.hardware.constants import RPI_HQ_CAMERA_CCM_MATRIX, RPI_HQ_CAMERA_BLACK_LEVEL
 
 try:
     import torch
@@ -237,33 +233,6 @@ def bayer2rgb(
     return (img * (2**nbits_out - 1)).astype(dtype)
 
 
-def get_distro():
-    """
-    Get current OS distribution.
-
-    Returns
-    -------
-    result : str
-        Name and version of OS.
-    """
-    # https://majornetwork.net/2019/11/get-linux-distribution-name-and-version-with-python/
-    RELEASE_DATA = {}
-    with open("/etc/os-release") as f:
-        reader = csv.reader(f, delimiter="=")
-        for row in reader:
-            if row:
-                RELEASE_DATA[row[0]] = row[1]
-    if RELEASE_DATA["ID"] in ["debian", "raspbian"]:
-        with open("/etc/debian_version") as f:
-            DEBIAN_VERSION = f.readline().strip()
-        major_version = DEBIAN_VERSION.split(".")[0]
-        version_split = RELEASE_DATA["VERSION"].split(" ", maxsplit=1)
-        if version_split[0] == major_version:
-            # Just major version shown, replace it with the full version
-            RELEASE_DATA["VERSION"] = " ".join([DEBIAN_VERSION] + version_split[1:])
-    return f"{RELEASE_DATA['NAME']} {RELEASE_DATA['VERSION']}"
-
-
 def print_image_info(img):
     """
     Print dimensions, data type, max, min, mean.
@@ -307,16 +276,3 @@ def autocorr2d(vals, pad_mode="reflect"):
 
     # remove padding
     return autocorr[shape[0] // 2 : -shape[0] // 2, shape[1] // 2 : -shape[1] // 2]
-
-
-def check_username_hostname(username, hostname, timeout=10):
-
-    client = paramiko.client.SSHClient()
-    client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-
-    try:
-        client.connect(hostname, username=username, timeout=timeout)
-    except (BadHostKeyException, AuthenticationException, SSHException, socket.error) as e:
-        raise ValueError(f"Could not connect to {username}@{hostname}\n{e}")
-
-    return username, hostname