+++ scripts/demo/telegram_bot.py
@@ -12,10 +12,11 @@ from PIL import Image, ImageFont
 import shutil
 import pytz
 from datetime import datetime
-from lensless.util import check_username_hostname
+from lensless.hardware.utils import check_username_hostname
 
 # for displaying emojis
 from emoji import EMOJI_DATA
+from lensless.utils.io import load_psf
 from pilmoji import Pilmoji
 
 from telegram import __version__ as TG_VER
@@ -874,7 +875,7 @@ def main(config) -> None:
 
     # load and downsample PSF beforehand
     if config.psf is not None:
-        from lensless.io import load_psf, save_image
+        from lensless.utils.io import save_image
 
         psf, bg = load_psf(
             config.psf.fp, downsample=config.psf.downsample, return_float=True, return_bg=True
@@ -885,7 +886,7 @@ def main(config) -> None:
         save_image(psf[0], PSF_FP)
 
         # save with gamma correction
-        from lensless.util import gamma_correction
+        from lensless.utils.image import gamma_correction
 
         psf_gamma = gamma_correction(psf[0], gamma=1.5)
         save_image(psf_gamma, PSF_FP_GAMMA)